# api/api-spec.yaml
openapi: 3.0.4
info:
  title: SIMBA API Server
  description: |
    ## SIMBA — Budget & Goods Management Information System

    Backend API for **Satuan Pelayanan Pemenuhan Gizi (SPPG)** — a nutrition fulfillment unit
    under the **Makan Bergizi Gratis (MBG)** free nutritious meals program.

    ### Core Features
    - **Authentication** — JWT-based login with refresh token rotation
    - **User Management** — Role-Based Access Control (RBAC) with Super Admin and Admin roles
    - **Item Management** — Inventory CRUD with stock tracking (IN/OUT transactions)
    - **Finance Management** — Expense/income recording with proof image upload
    - **Invoice Generation** — PDF invoice generation from stock transaction data
    - **Dashboard** — Aggregate analytics and monthly budget statistics
    - **Settings** — Company profile and document sequence configuration

    ### Authentication Flow
    1. `POST /api/auth/login` → receives `token` (HttpOnly cookie, 30min TTL)
       and `refresh_token` (HttpOnly cookie, 7-day TTL, scoped to `/api/auth`)
    2. All protected routes require valid `token` cookie
    3. `POST /api/auth/refresh` → rotates both cookies using the refresh token
    4. `DELETE /api/auth/logout` → clears both cookies

    ### Rate Limits
    - Global: **200 req/min** per IP
    - Auth endpoints: **10 req/min** per IP (brute-force protection)

    ### Roles
    | Role | Access |
    |------|--------|
    | `Super Admin` | Full access including user management |
...
servers:
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: |
        HttpOnly JWT cookie set automatically by `POST /api/auth/login`.
        Include credentials (`withCredentials: true`) in all API requests.

  # ─────────────────────────────────────────────
  # REUSABLE SCHEMAS
  # ─────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      description: Generic success response wrapper
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          description: Response payload (type varies per endpoint)
        paging:
          $ref: "#/components/schemas/PageMetadata"
        error:
          nullable: true
          example: null
    ErrorResponse:
      type: object
      description: Generic error response wrapper
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: "bad request"
        data:
          nullable: true
          example: null
        error:
          description: Error detail (string or validation error map)
    PageMetadata:
      type: object
      description: Pagination metadata included on list responses
      properties:
        page:
          type: integer
          description: Current page number (1-indexed)
          example: 1
        size:
          type: integer
          description: Items per page
          example: 10
        total_item:
          type: integer
          format: int64
          description: Total number of matching records
          example: 42
        total_page:
          type: integer
          format: int64
          description: Total number of pages
          example: 5

    # ── Auth ──────────────────────────────────
...
    CreateUserRequest:
      type: object
      required: [username, fullname, role, password]
      properties:
        username:
          type: string
          minLength: 5
          maxLength: 30
          pattern: "^[a-zA-Z0-9]+$"
          example: "johndoe"
        fullname:
          type: string
          minLength: 2
          maxLength: 50
          example: "John Doe"
        role:
          type: string
          enum: [Admin, "Super Admin"]
          example: "Admin"
        password:
          type: string
          minLength: 4
          maxLength: 30
          example: "pass1234"
    UpdateUserRequest:
      type: object
      properties:
        fullname:
          type: string
          minLength: 2
          maxLength: 50
          example: "John Doe Updated"
        role:
          type: string
          enum: [Admin, "Super Admin"]
          example: "Admin"
        password:
          type: string
          minLength: 4
          maxLength: 30
          example: "newPass123"
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "johndoe"
        fullname:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [Admin, "Super Admin"]
          example: "Admin"
        is_active:
          type: boolean
          description: True if user is currently logged in
          example: false
        last_active:
          type: string
          format: date-time
          nullable: true
          example: "2026-04-20T10:30:00Z"
        created_at:
          type: string
          format: date-time
          example: "2026-01-15T08:00:00Z"
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: Number of items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 401
            message: "user unauthorized"
            data: null
            error: "user unauthorized"
    Forbidden:
      description: Authenticated but role lacks permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 403
            message: "user forbidden"
            data: null
            error: "user forbidden"
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 404
            message: "not found"
            data: null
            error: "not found"
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 400
            message: "validation error"
            data: null
            error:
              username: "required"
              password: "min=4"
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List all users
      description: Returns a paginated list of users. **Super Admin only.**
      security:
        - cookieAuth: []
      parameters:
        - name: fullname
          in: query
          description: Filter by fullname (partial match)
          schema:
            type: string
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/UserResponse"
              example:
                status: 200
                message: "success"
                data:
                  - id: 1
                    username: "superadmin"
                    fullname: "Super Administrator"
                    role: "Super Admin"
                    is_active: true
                    last_active: "2026-04-21T10:00:00Z"
                    created_at: "2026-01-01T00:00:00Z"
                paging:
                  page: 1
                  size: 10
                  total_item: 1
                  total_page: 1
                error: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Users]
      summary: Create user
      description: Create a new user account. **Super Admin only.**
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Username already exists
          content:
            application/json:
              example:
                status: 409
                message: "user already exists"
                data: null
                error: "user already exists"
...
  /api/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: User ID (integer)
        schema:
          type: integer
          example: 1

    get:
      tags: [Users]
      summary: Get user by ID
      description: Returns a single user's details. **Super Admin only.**
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Users]
      summary: Update user
      description: Update user details. **Super Admin only.**
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Users]
      summary: Delete user
      description: |
        Soft-delete a user. **Super Admin only.**

        Restrictions:
        - Cannot delete yourself
        - Cannot delete the last Super Admin
        - Cannot delete a currently active (logged-in) user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              example:
                status: 200
                message: "user deleted"
                data: null
                error: null
        "400":
          description: Self-delete or business rule violation
          content:
            application/json:
              example:
                status: 400
                message: "cannot self delete user"
                data: null
                error: "cannot self delete user"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Items ─────────────────────────────────────────────────────────────────
...